lexibloom_-main/backend/requirements.txt
opencv-python
cvzone
pynput
pyperclip
pillow
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1